@@ -23,10 +23,9 @@
   "dependencies": {
     "frisbee": "^1.6.0",
     "history": "^4.7.2",
-    "netlify-identity-widget": "^1.0.1",
-    "prop-types": "^15.5.10",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
+    "prop-types": "^15.6.0",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
     "react-hot-loader": "next",
     "react-network": "^1.0.1",
     "react-redux": "^5.0.6",
@@ -51,7 +50,7 @@
     "eslint-plugin-jsx-a11y": "5.1",
     "eslint-plugin-react": "^7.3.0",
     "generact": "^0.3.1",
-    "nwb": "^0.18.10",
+    "nwb": "^0.19.1",
     "rimraf": "2.6.1"
   },
   "repository": {
