@@ -1,4 +1,5 @@
 import React from 'react'
+import PropTypes from 'prop-types'
 import { Provider } from 'react-redux'
 import { ConnectedRouter } from 'react-router-redux'
 import App from 'containers/App'
@@ -17,8 +18,8 @@ class Root extends React.Component {
 }
 
 Root.propTypes = {
-  history: React.PropTypes.object.isRequired,
-  store: React.PropTypes.object.isRequired,
+  history: PropTypes.object.isRequired,
+  store: PropTypes.object.isRequired,
 }
 
 export default Root
